save the python code in a seperate file and then run it
