+++ scripts/install.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 python3 -m venv venv
 source venv/bin/activate
-pip install -r requirements.txt
+pip3 install -r requirements.txt
 
 echo "installed all"
