+++ scripts/run.sh
@@ -1,4 +1,4 @@
 #!/bin/bash
 source venv/bin/activate
 cd src
-uvicorn start:app --reload
+uvicorn main:app --reload
